small refactoring of reactions/enumerators
